@@ -8,24 +8,33 @@ import (
 	"regexp"
 	"strings"
 
-	"github.com/ghodss/yaml"
-	"github.com/pkg/errors"
 	"github.com/xeipuuv/gojsonschema"
+	"sigs.k8s.io/yaml"
 
+	"github.com/akuity/kargo-render/internal/argocd"
 	"github.com/akuity/kargo-render/internal/file"
-	"github.com/akuity/kargo-render/internal/helm"
-	"github.com/akuity/kargo-render/internal/kustomize"
-	"github.com/akuity/kargo-render/internal/ytt"
 
 	_ "embed"
 )
 
 //go:embed schema.json
 var configSchemaBytes []byte
-var configSchemaJSONLoader gojsonschema.JSONLoader
+
+//go:embed argocd-schema.json
+var argocdConfigSchemaBytes []byte
+
+var configSchema *gojsonschema.Schema
 
 func init() {
-	configSchemaJSONLoader = gojsonschema.NewBytesLoader(configSchemaBytes)
+	sl := gojsonschema.NewSchemaLoader()
+	if err := sl.AddSchema("argocd-schema.json", gojsonschema.NewBytesLoader(argocdConfigSchemaBytes)); err != nil {
+		panic(fmt.Sprintf("error adding Argo CD schema: %s", err))
+	}
+
+	var err error
+	if configSchema, err = sl.Compile(gojsonschema.NewBytesLoader(configSchemaBytes)); err != nil {
+		panic(fmt.Sprintf("error compiling schema: %s", err))
+	}
 }
 
 // repoConfig encapsulates all Kargo Render configuration options for a
@@ -44,11 +53,11 @@ func (r *repoConfig) GetBranchConfig(name string) (branchConfig, error) {
 			regex, err := regexp.Compile(cfg.Pattern)
 			if err != nil {
 				return branchConfig{},
-					errors.Errorf("error compiling regular expression /%s/", cfg.Pattern)
+					fmt.Errorf("error compiling regular expression /%s/", cfg.Pattern)
 			}
 			submatches := regex.FindStringSubmatch(name)
 			if len(submatches) > 0 {
-				return cfg.expand(submatches), nil
+				return cfg.expand(submatches)
 			}
 		}
 	}
@@ -69,22 +78,42 @@ type branchConfig struct {
 	// PRs encapsulates details about how to manage any pull requests associated
 	// with this branch.
 	PRs pullRequestConfig `json:"prs,omitempty"`
+	// PreservedPaths specifies paths relative to the root of the repository that
+	// should be exempted from pre-render cleaning (deletion) of
+	// environment-specific branch contents. This is useful for preserving any
+	// environment-specific files that are manually maintained. Typically there
+	// are very few such files, if any at all, with an environment-specific
+	// CODEOWNERS file at the root of the repository being the most emblematic
+	// exception. Paths may be to files or directories. Any path to a directory
+	// will cause that directory's entire contents to be preserved.
+	PreservedPaths []string `json:"preservedPaths,omitempty"`
 }
 
-func (b branchConfig) expand(values []string) branchConfig {
+func (b branchConfig) expand(values []string) (branchConfig, error) {
 	cfg := b
 	cfg.AppConfigs = map[string]appConfig{}
 	for appName, appConfig := range b.AppConfigs {
-		cfg.AppConfigs[appName] = appConfig.expand(values)
+		var err error
+		if cfg.AppConfigs[appName], err = appConfig.expand(values); err != nil {
+			return cfg, fmt.Errorf(
+				"error expanding app config for app %q: %w",
+				appName,
+				err,
+			)
+		}
 	}
-	return cfg
+
+	for i, path := range b.PreservedPaths {
+		b.PreservedPaths[i] = file.ExpandPath(path, values)
+	}
+	return cfg, nil
 }
 
 // appConfig encapsulates application-specific Kargo Render configuration.
 type appConfig struct {
 	// ConfigManagement encapsulates configuration management options to be
 	// used with this branch and app.
-	ConfigManagement configManagementConfig `json:"configManagement,omitempty"`
+	ConfigManagement argocd.ConfigManagementConfig `json:"configManagement"`
 	// OutputPath specifies a path relative to the root of the repository where
 	// rendered manifests for this app will be stored in this branch.
 	OutputPath string `json:"outputPath,omitempty"`
@@ -93,40 +122,14 @@ type appConfig struct {
 	CombineManifests bool `json:"combineManifests,omitempty"`
 }
 
-func (a appConfig) expand(values []string) appConfig {
+func (a appConfig) expand(values []string) (appConfig, error) {
 	cfg := a
-	cfg.ConfigManagement = a.ConfigManagement.expand(values)
+	var err error
+	if cfg.ConfigManagement, err = a.ConfigManagement.Expand(values); err != nil {
+		return cfg, fmt.Errorf("error expanding config management config: %w", err)
+	}
 	cfg.OutputPath = file.ExpandPath(a.OutputPath, values)
-	return cfg
-}
-
-// configManagementConfig is a wrapper around more specific configuration for
-// one of three supported configuration management tools: helm, kustomize, or
-// ytt. Only one of its fields may be non-nil.
-type configManagementConfig struct { // nolint: revive
-	// Helm encapsulates optional Helm configuration options.
-	Helm *helm.Config `json:"helm,omitempty"`
-	// Kustomize encapsulates optional Kustomize configuration options.
-	Kustomize *kustomize.Config `json:"kustomize,omitempty"`
-	// Ytt encapsulates optional ytt configuration options.
-	Ytt *ytt.Config `json:"ytt,omitempty"`
-}
-
-func (c configManagementConfig) expand(values []string) configManagementConfig {
-	cfg := c
-	if c.Helm != nil {
-		helmCfg := c.Helm.Expand(values)
-		cfg.Helm = &helmCfg
-	}
-	if c.Kustomize != nil {
-		kustomizeCfg := c.Kustomize.Expand(values)
-		cfg.Kustomize = &kustomizeCfg
-	}
-	if c.Ytt != nil {
-		yttCfg := c.Ytt.Expand(values)
-		cfg.Ytt = &yttCfg
-	}
-	return cfg
+	return cfg, nil
 }
 
 // pullRequestConfig encapsulates details related to PR management for a branch.
@@ -164,12 +167,12 @@ func loadRepoConfig(repoPath string) (*repoConfig, error) {
 	var configPath string
 	if jsonExists, err := file.Exists(jsonConfigPath); err != nil {
 		return cfg,
-			errors.Wrap(err, "error checking for existence of JSON config file")
+			fmt.Errorf("error checking for existence of JSON config file: %w", err)
 	} else if jsonExists {
 		configPath = jsonConfigPath
 	} else if yamlExists, err := file.Exists(yamlConfigPath); err != nil {
 		return cfg,
-			errors.Wrap(err, "error checking for existence of YAML config file")
+			fmt.Errorf("error checking for existence of YAML config file: %w", err)
 	} else if yamlExists {
 		configPath = yamlConfigPath
 	}
@@ -178,16 +181,18 @@ func loadRepoConfig(repoPath string) (*repoConfig, error) {
 	}
 	configBytes, err := os.ReadFile(configPath)
 	if err != nil {
-		return cfg, errors.Wrap(err, "error reading Kargo Render configuration")
+		return cfg, fmt.Errorf("error reading Kargo Render configuration: %w", err)
 	}
 	if configBytes, err = normalizeAndValidate(configBytes); err != nil {
-		return cfg, errors.Wrap(
+		return cfg, fmt.Errorf(
+			"error normalizing and validating Kargo Render configuration: %w",
 			err,
-			"error normalizing and validating Kargo Render configuration",
 		)
 	}
-	err = json.Unmarshal(configBytes, cfg)
-	return cfg, errors.Wrap(err, "error unmarshaling Kargo Render configuration")
+	if err = json.Unmarshal(configBytes, cfg); err != nil {
+		return cfg, fmt.Errorf("error unmarshaling Kargo Render configuration: %w", err)
+	}
+	return cfg, nil
 }
 
 func normalizeAndValidate(configBytes []byte) ([]byte, error) {
@@ -196,21 +201,19 @@ func normalizeAndValidate(configBytes []byte) ([]byte, error) {
 	var err error
 	if configBytes, err = yaml.YAMLToJSON(configBytes); err != nil {
 		return nil,
-			errors.Wrap(err, "error normalizing Kargo Render configuration")
+			fmt.Errorf("error normalizing Kargo Render configuration: %w", err)
 	}
-	validationResult, err := gojsonschema.Validate(
-		configSchemaJSONLoader,
-		gojsonschema.NewBytesLoader(configBytes),
-	)
+
+	validationResult, err := configSchema.Validate(gojsonschema.NewBytesLoader(configBytes))
 	if err != nil {
-		return nil, errors.Wrap(err, "error validating Kargo Render configuration")
+		return nil, fmt.Errorf("error validating Kargo Render configuration: %w", err)
 	}
 	if !validationResult.Valid() {
 		verrStrs := make([]string, len(validationResult.Errors()))
 		for i, verr := range validationResult.Errors() {
 			verrStrs[i] = verr.String()
 		}
-		return nil, errors.Errorf(
+		return nil, fmt.Errorf(
 			"error validating Kargo Render configuration: %s",
 			strings.Join(verrStrs, "; "),
 		)
