@@ -10,9 +10,13 @@ import (
 	"path/filepath"
 	"strings"
 
-	"github.com/pkg/errors"
-
 	libExec "github.com/akuity/kargo-render/internal/exec"
+)
+
+const (
+	RemoteOrigin = "origin"
+
+	tmpPrefix = "repo-"
 )
 
 // RepoCredentials represents the credentials for connecting to a private git
@@ -47,7 +51,7 @@ type Repo interface {
 	// Checkout checks out the specified branch.
 	Checkout(branch string) error
 	// Commit commits staged changes to the current branch.
-	Commit(message string) error
+	Commit(message string, opts *CommitOptions) error
 	// CreateChildBranch creates a new branch that is a child of the current
 	// branch.
 	CreateChildBranch(branch string) error
@@ -64,17 +68,28 @@ type Repo interface {
 	// LastCommitID returns the ID (sha) of the most recent commit to the current
 	// branch.
 	LastCommitID() (string, error)
+	// LocalBranchExists returns a bool indicating if the specified branch exists.
+	LocalBranchExists(branch string) (bool, error)
 	// CommitMessage returns the text of the most recent commit message associated
 	// with the specified commit ID.
 	CommitMessage(id string) (string, error)
 	// CommitMessages returns a slice of commit messages starting with id1 and
 	// ending with id2. The results exclude id1, but include id2.
 	CommitMessages(id1, id2 string) ([]string, error)
+	// Fetch fetches from the remote repository.
+	Fetch() error
+	// Pull fetches from the remote repository and merges the changes into the
+	// current branch.
+	Pull(branch string) error
 	// Push pushes from the current branch to a remote branch by the same name.
 	Push() error
 	// RemoteBranchExists returns a bool indicating if the specified branch exists
 	// in the remote repository.
 	RemoteBranchExists(branch string) (bool, error)
+	// Remotes returns a slice of strings representing the names of the remotes.
+	Remotes() ([]string, error)
+	// RemoteURL returns the URL of the the specified remote.
+	RemoteURL(name string) (string, error)
 	// ResetHard performs a hard reset.
 	ResetHard() error
 	// URL returns the remote URL of the repository.
@@ -93,6 +108,7 @@ type repo struct {
 	homeDir       string
 	dir           string
 	currentBranch string
+	creds         RepoCredentials
 }
 
 // Clone produces a local clone of the remote git repository at the specified
@@ -101,21 +117,22 @@ type repo struct {
 // perform any setup that is required for successfully authenticating to the
 // remote repository.
 func Clone(
-	url string,
+	cloneURL string,
 	repoCreds RepoCredentials,
 ) (Repo, error) {
-	homeDir, err := os.MkdirTemp("", "")
+	homeDir, err := os.MkdirTemp("", tmpPrefix)
 	if err != nil {
-		return nil, errors.Wrapf(
+		return nil, fmt.Errorf(
+			"error creating home directory for repo %q: %w",
+			cloneURL,
 			err,
-			"error creating home directory for repo %q",
-			url,
 		)
 	}
 	r := &repo{
-		url:     url,
+		url:     cloneURL,
 		homeDir: homeDir,
 		dir:     filepath.Join(homeDir, "repo"),
+		creds:   repoCreds,
 	}
 	if err = r.setupAuth(repoCreds); err != nil {
 		return nil, err
@@ -123,34 +140,114 @@ func Clone(
 	return r, r.clone()
 }
 
+// CopyRepo copies a git repository from the specified path to a temporary
+// location. Repository credentials are required in order to authenticate to the
+// remote repository, if any.
+func CopyRepo(path string, repoCreds RepoCredentials) (Repo, error) {
+	// Validate path is absolute
+	if !filepath.IsAbs(path) {
+		return nil, fmt.Errorf("path %s is not absolute", path)
+	}
+
+	// Validate path exists and is a directory
+	if fi, err := os.Stat(path); err != nil {
+		return nil, fmt.Errorf("error checking if path %s exists: %w", path, err)
+	} else if !fi.IsDir() {
+		return nil, fmt.Errorf("path %s is not a directory", path)
+	}
+
+	// Validate path is a git repository
+	cmd := exec.Command("git", "rev-parse", "--is-inside-work-tree")
+	cmd.Dir = path
+	if _, err := libExec.Exec(cmd); err != nil {
+		return nil, fmt.Errorf("path %s is not a git repository: %w", path, err)
+	}
+
+	homeDir, err := os.MkdirTemp("", tmpPrefix)
+	if err != nil {
+		return nil, fmt.Errorf(
+			"error creating directory for copy of repo at %s: %w",
+			path,
+			err,
+		)
+	}
+
+	r := &repo{
+		homeDir: homeDir,
+		dir:     filepath.Join(homeDir, "repo"),
+	}
+
+	// Copy from path to r.dir. Note: This obviously only works on *nix systems,
+	// but we already advise that Kargo Render not be run outside of a Linux
+	// container since its dependent on compatible versions of git, helm, and
+	// kustomize binaries.
+	// nolint: gosec
+	if _, err = libExec.Exec(exec.Command("cp", "-r", path, r.dir)); err != nil {
+		return nil, fmt.Errorf(
+			"error copying repo from %s to %s: %w",
+			path,
+			r.dir,
+			err,
+		)
+	}
+
+	remotes, err := r.Remotes()
+	if err != nil {
+		return nil, err
+	}
+	if len(remotes) != 1 {
+		return nil, fmt.Errorf(
+			"expected exactly one remote in source repository; found %d",
+			len(remotes),
+		)
+	}
+	r.url, err = r.RemoteURL(remotes[0])
+	if err != nil {
+		return nil, err
+	}
+
+	if err = r.setupAuth(repoCreds); err != nil {
+		return nil, err
+	}
+
+	return r, nil
+}
+
 func (r *repo) AddAll() error {
-	_, err := libExec.Exec(r.buildCommand("add", "."))
-	return errors.Wrap(err, "error staging changes for commit")
+	if _, err := libExec.Exec(r.buildCommand("add", ".")); err != nil {
+		return fmt.Errorf("error staging changes for commit: %w", err)
+	}
+	return nil
 }
 
 func (r *repo) AddAllAndCommit(message string) error {
 	if err := r.AddAll(); err != nil {
 		return err
 	}
-	return r.Commit(message)
+	return r.Commit(message, nil)
 }
 
 func (r *repo) Clean() error {
 	_, err := libExec.Exec(r.buildCommand("clean", "-fd"))
-	return errors.Wrapf(err, "error cleaning branch %q", r.currentBranch)
+	if err != nil {
+		return fmt.Errorf("error cleaning branch %q: %w", r.currentBranch, err)
+	}
+	return nil
 }
 
 func (r *repo) clone() error {
 	r.currentBranch = "HEAD"
 	cmd := r.buildCommand("clone", "--no-tags", r.url, r.dir)
 	cmd.Dir = r.homeDir // Override the cmd.Dir that's set by r.buildCommand()
-	_, err := libExec.Exec(cmd)
-	return errors.Wrapf(
-		err,
-		"error cloning repo %q into %q",
-		r.url,
-		r.dir,
-	)
+	if _, err := libExec.Exec(cmd); err != nil {
+		return fmt.Errorf(
+			"error cloning repo %q into %q: %w",
+			r.url,
+			r.dir,
+			err,
+		)
+	}
+	return nil
 }
 
 func (r *repo) Close() error {
@@ -159,34 +256,49 @@ func (r *repo) Close() error {
 
 func (r *repo) Checkout(branch string) error {
 	r.currentBranch = branch
-	_, err := libExec.Exec(r.buildCommand(
+	if _, err := libExec.Exec(r.buildCommand(
 		"checkout",
 		branch,
 		// The next line makes it crystal clear to git that we're checking out
 		// a branch. We need to do this because branch names can often resemble
 		// paths within the repo.
 		"--",
-	))
-	return errors.Wrapf(
-		err,
-		"error checking out branch %q from repo %q",
-		branch,
-		r.url,
-	)
+	)); err != nil {
+		return fmt.Errorf(
+			"error checking out branch %q from repo %q: %w",
+			branch,
+			r.url,
+			err,
+		)
+	}
+	return nil
 }
 
-func (r *repo) Commit(message string) error {
-	_, err := libExec.Exec(r.buildCommand("commit", "-m", message))
-	return errors.Wrapf(
-		err,
-		"error committing changes to branch %q",
-		r.currentBranch,
-	)
+type CommitOptions struct {
+	AllowEmpty bool
+}
+
+func (r *repo) Commit(message string, opts *CommitOptions) error {
+	if opts == nil {
+		opts = &CommitOptions{}
+	}
+	cmdTokens := []string{"commit", "-m", message}
+	if opts.AllowEmpty {
+		cmdTokens = append(cmdTokens, "--allow-empty")
+	}
+	if _, err := libExec.Exec(r.buildCommand(cmdTokens...)); err != nil {
+		return fmt.Errorf(
+			"error committing changes to branch %q: %w",
+			r.currentBranch,
+			err,
+		)
+	}
+	return nil
 }
 
 func (r *repo) CreateChildBranch(branch string) error {
 	r.currentBranch = branch
-	_, err := libExec.Exec(r.buildCommand(
+	if _, err := libExec.Exec(r.buildCommand(
 		"checkout",
 		"-b",
 		branch,
@@ -194,13 +306,15 @@ func (r *repo) CreateChildBranch(branch string) error {
 		// a branch. We need to do this because branch names can often resemble
 		// paths within the repo.
 		"--",
-	))
-	return errors.Wrapf(
-		err,
-		"error creating new branch %q for repo %q",
-		branch,
-		r.url,
-	)
+	)); err != nil {
+		return fmt.Errorf(
+			"error creating new branch %q for repo %q: %w",
+			branch,
+			r.url,
+			err,
+		)
+	}
+	return nil
 }
 
 func (r *repo) CreateOrphanedBranch(branch string) error {
@@ -211,11 +325,11 @@ func (r *repo) CreateOrphanedBranch(branch string) error {
 		branch,
 		"--discard-changes",
 	)); err != nil {
-		return errors.Wrapf(
-			err,
-			"error creating orphaned branch %q for repo %q",
+		return fmt.Errorf(
+			"error creating orphaned branch %q for repo %q: %w",
 			branch,
 			r.url,
+			err,
 		)
 	}
 	return r.Clean()
@@ -223,15 +337,18 @@ func (r *repo) CreateOrphanedBranch(branch string) error {
 
 func (r *repo) HasDiffs() (bool, error) {
 	resBytes, err := libExec.Exec(r.buildCommand("status", "-s"))
-	return len(resBytes) > 0,
-		errors.Wrapf(err, "error checking status of branch %q", r.currentBranch)
+	if err != nil {
+		return false,
+			fmt.Errorf("error checking status of branch %q: %w", r.currentBranch, err)
+	}
+	return len(resBytes) > 0, nil
 }
 
 func (r *repo) GetDiffPaths() ([]string, error) {
 	resBytes, err := libExec.Exec(r.buildCommand("status", "-s"))
 	if err != nil {
 		return nil,
-			errors.Wrapf(err, "error checking status of branch %q", r.currentBranch)
+			fmt.Errorf("error checking status of branch %q: %w", r.currentBranch, err)
 	}
 	paths := []string{}
 	scanner := bufio.NewScanner(bytes.NewReader(resBytes))
@@ -247,16 +364,36 @@ func (r *repo) GetDiffPaths() ([]string, error) {
 
 func (r *repo) LastCommitID() (string, error) {
 	shaBytes, err := libExec.Exec(r.buildCommand("rev-parse", "HEAD"))
-	return strings.TrimSpace(string(shaBytes)),
-		errors.Wrap(err, "error obtaining ID of last commit")
+	if err != nil {
+		return "", fmt.Errorf("error obtaining ID of last commit: %w", err)
+	}
+	return strings.TrimSpace(string(shaBytes)), nil
+}
+
+func (r *repo) LocalBranchExists(branch string) (bool, error) {
+	resBytes, err := libExec.Exec(r.buildCommand(
+		"branch",
+		"--list",
+		branch,
+	))
+	if err != nil {
+		return false,
+			fmt.Errorf("error checking for existence of local branch %q: %w", branch, err)
+	}
+	return strings.TrimSpace(
+		strings.Replace(string(resBytes), "*", "", 1),
+	) == branch, nil
 }
 
 func (r *repo) CommitMessage(id string) (string, error) {
 	msgBytes, err := libExec.Exec(
 		r.buildCommand("log", "-n", "1", "--pretty=format:%s", id),
 	)
-	return string(msgBytes),
-		errors.Wrapf(err, "error obtaining commit message for commit %q", id)
+	if err != nil {
+		return "",
+			fmt.Errorf("error obtaining commit message for commit %q: %w", id, err)
+	}
+	return string(msgBytes), nil
 }
 
 func (r *repo) CommitMessages(id1, id2 string) ([]string, error) {
@@ -268,11 +405,11 @@ func (r *repo) CommitMessages(id1, id2 string) ([]string, error) {
 		fmt.Sprintf("%s..%s", id1, id2),
 	))
 	if err != nil {
-		return nil, errors.Wrapf(
-			err,
-			"error obtaining commit messages between commits %q and %q",
+		return nil, fmt.Errorf(
+			"error obtaining commit messages between commits %q and %q: %w",
 			id1,
 			id2,
+			err,
 		)
 	}
 	msgsBytes := bytes.Split(allMsgBytes, []byte("\n"))
@@ -289,36 +426,83 @@ func (r *repo) CommitMessages(id1, id2 string) ([]string, error) {
 	return msgs, nil
 }
 
+func (r *repo) Fetch() error {
+	if _, err := libExec.Exec(r.buildCommand("fetch", RemoteOrigin)); err != nil {
+		return fmt.Errorf("error fetching from remote repo %q: %w", r.url, err)
+	}
+	return nil
+}
+
+func (r *repo) Pull(branch string) error {
+	if _, err :=
+		libExec.Exec(r.buildCommand("pull", RemoteOrigin, branch)); err != nil {
+		return fmt.Errorf(
+			"error pulling branch %q from remote repo %q: %w",
+			branch,
+			r.url,
+			err,
+		)
+	}
+	return nil
+}
+
 func (r *repo) Push() error {
-	_, err :=
-		libExec.Exec(r.buildCommand("push", "origin", r.currentBranch))
-	return errors.Wrapf(err, "error pushing branch %q", r.currentBranch)
+	if _, err :=
+		libExec.Exec(r.buildCommand("push", RemoteOrigin, r.currentBranch)); err != nil {
+		return fmt.Errorf("error pushing branch %q: %w", r.currentBranch, err)
+	}
+	return nil
 }
 
 func (r *repo) RemoteBranchExists(branch string) (bool, error) {
-	_, err := libExec.Exec(r.buildCommand(
+	if _, err := libExec.Exec(r.buildCommand(
 		"ls-remote",
 		"--heads",
 		"--exit-code", // Return 2 if not found
-		r.url,
+		RemoteOrigin,
 		branch,
-	))
-	if exitErr, ok := err.(*libExec.ExitError); ok && exitErr.ExitCode == 2 {
-		// Branch does not exist
-		return false, nil
+	)); err != nil {
+		if exitErr, ok := err.(*libExec.ExitError); ok && exitErr.ExitCode == 2 {
+			// Branch does not exist
+			return false, nil
+		}
+		return false, fmt.Errorf(
+			"error checking for existence of branch %q in remote repo %q: %w",
+			branch,
+			r.url,
+			err,
+		)
 	}
-	return err == nil, errors.Wrapf(
-		err,
-		"error checking for existence of branch %q in remote repo %q",
-		branch,
-		r.url,
-	)
+	return true, nil
+}
+
+func (r *repo) Remotes() ([]string, error) {
+	resBytes, err := libExec.Exec(r.buildCommand("remote"))
+	if err != nil {
+		return nil, fmt.Errorf("error listing remotes for repo %q: %w", r.url, err)
+	}
+	return strings.Fields(string(resBytes)), nil
+}
+
+func (r *repo) RemoteURL(name string) (string, error) {
+	resBytes, err := libExec.Exec(r.buildCommand("remote", "get-url", name))
+	if err != nil {
+		return "", fmt.Errorf(
+			"error obtaining URL for remote %q of repo %q: %w",
+			name,
+			r.url,
+			err,
+		)
+	}
+	return strings.TrimSpace(string(resBytes)), nil
 }
 
 func (r *repo) ResetHard() error {
-	_, err :=
-		libExec.Exec(r.buildCommand("reset", "--hard"))
-	return errors.Wrap(err, "error resetting branch working tree")
+	if _, err :=
+		libExec.Exec(r.buildCommand("reset", "--hard")); err != nil {
+		return fmt.Errorf("error resetting branch working tree: %w", err)
+	}
+	return nil
 }
 
 func (r *repo) URL() string {
@@ -340,13 +524,13 @@ func (r *repo) setupAuth(repoCreds RepoCredentials) error {
 	cmd := r.buildCommand("config", "--global", "user.name", "Kargo Render")
 	cmd.Dir = r.homeDir // Override the cmd.Dir that's set by r.buildCommand()
 	if _, err := libExec.Exec(cmd); err != nil {
-		return errors.Wrapf(err, "error configuring git username")
+		return fmt.Errorf("error configuring git username: %w", err)
 	}
 	cmd =
 		r.buildCommand("config", "--global", "user.email", "kargo-render@akuity.io")
 	cmd.Dir = r.homeDir // Override the cmd.Dir that's set by r.buildCommand()
 	if _, err := libExec.Exec(cmd); err != nil {
-		return errors.Wrapf(err, "error configuring git user email address")
+		return fmt.Errorf("error configuring git user email address: %w", err)
 	}
 
 	// If an SSH key was provided, use that.
@@ -356,7 +540,7 @@ func (r *repo) setupAuth(repoCreds RepoCredentials) error {
 		const sshConfig = "Host *\n  StrictHostKeyChecking no\n  UserKnownHostsFile=/dev/null"
 		if err :=
 			os.WriteFile(sshConfigPath, []byte(sshConfig), 0600); err != nil {
-			return errors.Wrapf(err, "error writing SSH config to %q", sshConfigPath)
+			return fmt.Errorf("error writing SSH config to %q: %w", sshConfigPath, err)
 		}
 
 		rsaKeyPath := filepath.Join(r.homeDir, ".ssh", "id_rsa")
@@ -365,48 +549,25 @@ func (r *repo) setupAuth(repoCreds RepoCredentials) error {
 			[]byte(repoCreds.SSHPrivateKey),
 			0600,
 		); err != nil {
-			return errors.Wrapf(err, "error writing SSH key to %q", rsaKeyPath)
+			return fmt.Errorf("error writing SSH key to %q: %w", rsaKeyPath, err)
 		}
 		return nil // We're done
 	}
 
-	// If we get to here, we're authenticating using a password
-
-	// Set up the credential helper
-	cmd = r.buildCommand("config", "--global", "credential.helper", "store")
-	cmd.Dir = r.homeDir // Override the cmd.Dir that's set by r.buildCommand()
-	if _, err := libExec.Exec(cmd); err != nil {
-		return errors.Wrapf(err, "error configuring git credential helper")
+	// If no password is specified, we're done'.
+	if repoCreds.Password == "" {
+		return nil
 	}
 
-	credentialURL, err := url.Parse(r.url)
-	if err != nil {
-		return errors.Wrapf(err, "error parsing URL %q", r.url)
-	}
-	// Remove path and query string components from the URL
-	credentialURL.Path = ""
-	credentialURL.RawQuery = ""
-	// If the username is the empty string, we assume we're working with a git
-	// provider like GitHub that only requires the username to be non-empty. We
-	// arbitrarily set it to "git".
-	if repoCreds.Username == "" {
-		repoCreds.Username = "git"
-	}
-	// Augment the URL with user/pass information.
-	credentialURL.User = url.UserPassword(repoCreds.Username, repoCreds.Password)
-	// Write the augmented URL to the location used by the "stored" credential
-	// helper.
-	credentialsPath := filepath.Join(r.homeDir, ".git-credentials")
-	if err := os.WriteFile(
-		credentialsPath,
-		[]byte(credentialURL.String()),
-		0600,
-	); err != nil {
-		return errors.Wrapf(
-			err,
-			"error writing credentials to %q",
-			credentialsPath,
-		)
+	lowerURL := strings.ToLower(r.url)
+	if strings.HasPrefix(lowerURL, "http://") || strings.HasPrefix(lowerURL, "https://") {
+		u, err := url.Parse(r.url)
+		if err != nil {
+			return fmt.Errorf("error parsing URL %q: %w", r.url, err)
+		}
+		u.User = url.User(repoCreds.Username)
+		r.url = u.String()
+
 	}
 	return nil
 }
@@ -418,6 +579,13 @@ func (r *repo) buildCommand(arg ...string) *exec.Cmd {
 		cmd.Env = []string{homeEnvVar}
 	} else {
 		cmd.Env = append(cmd.Env, homeEnvVar)
+	}
+	if r.creds.Password != "" {
+		cmd.Env = append(
+			cmd.Env,
+			"GIT_ASKPASS=/usr/local/bin/credential-helper",
+			fmt.Sprintf("GIT_PASSWORD=%s", r.creds.Password),
+		)
 	}
 	cmd.Dir = r.dir
 	return cmd
