@@ -6,9 +6,6 @@ import (
 	"os"
 	"path/filepath"
 
-	"github.com/pkg/errors"
-	log "github.com/sirupsen/logrus"
-
 	"github.com/akuity/kargo-render/internal/kustomize"
 	"github.com/akuity/kargo-render/internal/strings"
 )
@@ -25,73 +22,18 @@ resources:
 func (s *service) preRender(
 	ctx context.Context,
 	rc requestContext,
-	repoDir string,
+	repoRoot string,
 ) (map[string][]byte, error) {
 	logger := rc.logger
 	manifests := map[string][]byte{}
 	var err error
 	for appName, appConfig := range rc.target.branchConfig.AppConfigs {
 		appLogger := logger.WithField("app", appName)
-		if appConfig.ConfigManagement.Helm != nil {
-			chartPath := appConfig.ConfigManagement.Helm.ChartPath
-			if chartPath == "" {
-				chartPath = "base"
-			}
-			valuesPaths := appConfig.ConfigManagement.Helm.ValuesPaths
-			if len(valuesPaths) == 0 {
-				valuesPaths =
-					[]string{filepath.Join(rc.request.TargetBranch, "values.yaml")}
-			}
-			appLogger = appLogger.WithFields(log.Fields{
-				"configManagement": "helm",
-				"releaseName":      appConfig.ConfigManagement.Helm.ReleaseName,
-				"namespace":        appConfig.ConfigManagement.Helm.Namespace,
-				"chartPath":        chartPath,
-				"valuesPaths":      valuesPaths,
-			})
-			chartPath = filepath.Join(repoDir, chartPath)
-			absValuesPaths := make([]string, len(valuesPaths))
-			for i, valuesPath := range valuesPaths {
-				absValuesPaths[i] = filepath.Join(repoDir, valuesPath)
-			}
-			manifests[appName], err = s.helmRenderFn(
-				ctx,
-				appConfig.ConfigManagement.Helm.ReleaseName,
-				appConfig.ConfigManagement.Helm.Namespace,
-				chartPath,
-				absValuesPaths,
-			)
-		} else if appConfig.ConfigManagement.Ytt != nil {
-			paths := appConfig.ConfigManagement.Ytt.Paths
-			if len(paths) == 0 {
-				paths = []string{"base", rc.request.TargetBranch}
-			}
-			appLogger = appLogger.WithFields(log.Fields{
-				"configManagement": "ytt",
-				"paths":            paths,
-			})
-			absPaths := make([]string, len(paths))
-			for i, path := range paths {
-				absPaths[i] = filepath.Join(repoDir, path)
-			}
-			manifests[appName], err = s.yttRenderFn(ctx, absPaths)
-		} else {
-			path := appConfig.ConfigManagement.Kustomize.Path
-			if path == "" {
-				path = rc.request.TargetBranch
-			}
-			appLogger = appLogger.WithFields(log.Fields{
-				"configManagement": "kustomize",
-				"path":             path,
-			})
-			path = filepath.Join(repoDir, path)
-			manifests[appName], err = s.kustomizeRenderFn(
-				ctx,
-				path,
-				nil,
-				*appConfig.ConfigManagement.Kustomize,
-			)
-		}
+		manifests[appName], err = s.renderFn(
+			ctx,
+			repoRoot,
+			appConfig.ConfigManagement,
+		)
 		if err != nil {
 			return nil, err
 		}
@@ -102,7 +44,7 @@ func (s *service) preRender(
 		// This is a sanity check. Argo CD does this also.
 		for appName := range rc.target.branchConfig.AppConfigs {
 			if manifests, ok := manifests[appName]; !ok || len(manifests) == 0 {
-				return nil, errors.Errorf(
+				return nil, fmt.Errorf(
 					"pre-rendered manifests for app %q contain 0 bytes; this looks "+
 						"like a mistake and allowEmpty is not set; refusing to proceed",
 					appName,
@@ -119,12 +61,12 @@ func renderLastMile(
 ) ([]string, map[string][]byte, error) {
 	logger := rc.logger
 
-	tempDir, err := os.MkdirTemp("", "")
+	tempDir, err := os.MkdirTemp("", "repo-scrap-")
 	if err != nil {
-		return nil, nil, errors.Wrapf(
-			err,
-			"error creating temporary directory %q for last mile rendering",
+		return nil, nil, fmt.Errorf(
+			"error creating temporary directory %q for last mile rendering: %w",
 			tempDir,
+			err,
 		)
 	}
 	defer os.RemoveAll(tempDir)
@@ -161,11 +103,11 @@ func renderLastMile(
 	for appName := range rc.target.branchConfig.AppConfigs {
 		appDir := filepath.Join(tempDir, appName)
 		if err = os.MkdirAll(appDir, 0755); err != nil {
-			return nil, nil, errors.Wrapf(
-				err,
-				"error creating directory %q for last mile rendering of app %q",
+			return nil, nil, fmt.Errorf(
+				"error creating directory %q for last mile rendering of app %q: %w",
 				appDir,
 				appName,
+				err,
 			)
 		}
 		// Create kustomization.yaml
@@ -175,10 +117,10 @@ func renderLastMile(
 			lastMileKustomizationBytes,
 			0644,
 		); err != nil {
-			return nil, nil, errors.Wrapf(
-				err,
-				"error writing last-mile kustomization.yaml to %q",
+			return nil, nil, fmt.Errorf(
+				"error writing last-mile kustomization.yaml to %q: %w",
 				appKustomizationFile,
+				err,
 			)
 		}
 		// Write the pre-rendered manifests to a file
@@ -189,18 +131,18 @@ func renderLastMile(
 			rc.target.prerenderedManifests[appName],
 			0644,
 		); err != nil {
-			return nil, nil, errors.Wrapf(
-				err,
-				"error writing pre-rendered manifests to %q",
+			return nil, nil, fmt.Errorf(
+				"error writing pre-rendered manifests to %q: %w",
 				preRenderedPath,
+				err,
 			)
 		}
 		if manifests[appName], err =
-			kustomize.Render(ctx, appDir, images, kustomize.Config{}); err != nil {
-			return nil, nil, errors.Wrapf(
-				err,
-				"error rendering manifests from %q",
+			kustomize.Render(ctx, appDir, images); err != nil {
+			return nil, nil, fmt.Errorf(
+				"error rendering manifests from %q: %w",
 				appDir,
+				err,
 			)
 		}
 		logger.WithField("app", appName).
