@@ -10,41 +10,64 @@ func TestValidateAndCanonicalizeRequest(t *testing.T) {
 	testCases := []struct {
 		name       string
 		req        Request
-		assertions func(Request, error)
+		assertions func(*testing.T, Request, error)
 	}{
 		{
-			name: "missing RepoURL",
+			name: "no input source specified",
 			req:  Request{},
-			assertions: func(req Request, err error) {
+			assertions: func(t *testing.T, _ Request, err error) {
 				require.Error(t, err)
-				require.Contains(t, err.Error(), "RepoURL is a required field")
+				require.Contains(t, err.Error(), "no input source specified")
+			},
+		},
+		{
+			name: "input source is ambiguous",
+			req: Request{
+				RepoURL:     "fake-url",
+				LocalInPath: "/some/path",
+			},
+			assertions: func(t *testing.T, _ Request, err error) {
+				require.Error(t, err)
+				require.Contains(t, err.Error(), "input source is ambiguous")
+			},
+		},
+		{
+			name: "local input path and git ref incorrectly used together",
+			req: Request{
+				LocalInPath: "/some/path",
+				Ref:         "1abcdef2",
+			},
+			assertions: func(t *testing.T, _ Request, err error) {
+				require.Error(t, err)
+				require.Contains(
+					t,
+					err.Error(),
+					"LocalInPath and Ref are mutually exclusive",
+				)
+			},
+		},
+		{
+			name: "output destination is ambiguous",
+			req: Request{
+				LocalOutPath: "/some/path",
+				Stdout:       true,
+			},
+			assertions: func(t *testing.T, _ Request, err error) {
+				require.Error(t, err)
+				require.Contains(t, err.Error(), "output destination is ambiguous")
 			},
 		},
 		{
 			name: "invalid RepoURL",
 			req: Request{
-				RepoURL: "foobar",
+				RepoURL: "fake-url",
 			},
-			assertions: func(req Request, err error) {
+			assertions: func(t *testing.T, _ Request, err error) {
 				require.Error(t, err)
 				require.Contains(
 					t,
 					err.Error(),
 					"does not appear to be a valid git repository URL",
-				)
-			},
-		},
-		{
-			name: "missing Password",
-			req: Request{
-				RepoURL: "https://github.com/akuity/foobar",
-			},
-			assertions: func(req Request, err error) {
-				require.Error(t, err)
-				require.Contains(
-					t,
-					err.Error(),
-					"RepoCreds.Password is a required field",
 				)
 			},
 		},
@@ -57,7 +80,7 @@ func TestValidateAndCanonicalizeRequest(t *testing.T) {
 				},
 				Ref: "1abcdef2",
 			},
-			assertions: func(req Request, err error) {
+			assertions: func(t *testing.T, _ Request, err error) {
 				require.Error(t, err)
 				require.Contains(t, err.Error(), "TargetBranch is a required field")
 			},
@@ -72,7 +95,7 @@ func TestValidateAndCanonicalizeRequest(t *testing.T) {
 				Ref:          "1abcdef2",
 				TargetBranch: "env/dev*", // * is an invalid character
 			},
-			assertions: func(req Request, err error) {
+			assertions: func(t *testing.T, _ Request, err error) {
 				require.Error(t, err)
 				require.Contains(t, err.Error(), "is an invalid branch name")
 			},
@@ -88,13 +111,33 @@ func TestValidateAndCanonicalizeRequest(t *testing.T) {
 				TargetBranch: "env/dev",
 				Images:       []string{""}, // no good
 			},
-			assertions: func(req Request, err error) {
+			assertions: func(t *testing.T, _ Request, err error) {
 				require.Error(t, err)
 				require.Contains(
 					t,
 					err.Error(),
 					"Images must not contain any empty strings",
 				)
+			},
+		},
+		{
+			name: "LocalInPath does not exist",
+			req: Request{
+				LocalInPath: "/some/path/that/does/not/exist",
+			},
+			assertions: func(t *testing.T, _ Request, err error) {
+				require.Error(t, err)
+				require.Contains(t, err.Error(), "does not exist")
+			},
+		},
+		{
+			name: "LocalOutPath exists",
+			req: Request{
+				LocalOutPath: t.TempDir(),
+			},
+			assertions: func(t *testing.T, _ Request, err error) {
+				require.Error(t, err)
+				require.Contains(t, err.Error(), "already exists; refusing to overwrite")
 			},
 		},
 		{
@@ -108,7 +151,7 @@ func TestValidateAndCanonicalizeRequest(t *testing.T) {
 				TargetBranch: "  refs/heads/env/dev  ",
 				Images:       []string{" akuity/some-image "}, // no good
 			},
-			assertions: func(req Request, err error) {
+			assertions: func(t *testing.T, req Request, err error) {
 				require.NoError(t, err)
 				require.Equal(t, "https://github.com/akuity/foobar", req.RepoURL)
 				require.Equal(t, "foobar", req.RepoCreds.Password)
@@ -120,7 +163,8 @@ func TestValidateAndCanonicalizeRequest(t *testing.T) {
 	}
 	for _, testCase := range testCases {
 		t.Run(testCase.name, func(t *testing.T) {
-			testCase.assertions(validateAndCanonicalizeRequest(testCase.req))
+			err := testCase.req.canonicalizeAndValidate()
+			testCase.assertions(t, testCase.req, err)
 		})
 	}
 }
